feat [#82] Group, Member type 선언
