+++ src/features/group/apis/index.ts
@@ -1,5 +1,5 @@
 import { apiRequest, type APIRequestOptions } from "@/services/api-request";
-import { UserGroup } from "@/types/user-group";
+import { UserGroup } from "@/types/group";
 
 export async function getUserGroups(options?: APIRequestOptions) {
   return apiRequest<UserGroup[]>(async (instance) => {
