+++ src/types/member.ts
@@ -1,0 +1,10 @@
+export interface Member {
+  role: Role;
+  userImage: string;
+  userEmail: string;
+  userName: string;
+  groupId: number;
+  userId: number;
+}
+
+export type Role = "ADMIN" | "MEMBER";
