--- src/types/user-group.ts
@@ -1,8 +1,0 @@
-export interface UserGroup {
-  teamId: string;
-  updatedAt: string;
-  createdAt: string;
-  image: string;
-  name: string;
-  id: number;
-}
